==> 0-Bootstraping.ipynb <==
from tqdm.auto import tqdm
import random, json

# setup seed to reproduce results
random.seed(42)

with open("data/data-for-trf.json", "r", encoding="utf-8") as f:
    idsselected = []
    print("loading json...")
    jsonfile = json.load(f)
    print("json loaded")

    # randomly draw data from raw dataset to build train dataset
    data = []
    while len(data) != len(jsonfile):
        datum = random.choice(jsonfile)
        if not (datum["id"] in idsselected):
            idsselected.append(datum["id"])
        data.append(datum)
    with open("data/data-for-trf-train.json", "w", encoding="utf-8") as trainfile:
        json.dump(data, trainfile, ensure_ascii=False, indent=4)
    print("Number of data in train dataset (without counting duplicates):", len(idsselected))

    # get data not included into the train dataset to build validation dataset
    data = [datum for datum in tqdm(jsonfile) if not (datum["id"] in idsselected)]
    with open("data/data-for-trf-validation.json", "w", encoding="utf-8") as validfile:
        json.dump(data, validfile, ensure_ascii=False, indent=4)
    print("Number of data in validation dataset:", len(data))
